starknet_native_compile

Function load_sierra_program_from_file

pub(crate) fn load_sierra_program_from_file(
    path: &PathBuf,
) -> Result<(ContractClass, Program, VersionId)>

Given a Sierra file path, loads the contract class from it, extracts the sierra version from the first 3 felts of the compressed sierra_program, and extracts the compressed sierra_program into readable sierra code.